+++ Android/src/org/droidplanner/android/fragments/DroneMap.java
@@ -17,7 +17,13 @@ import org.droidplanner.android.utils.prefs.AutoPanMode;
 import org.droidplanner.core.drone.DroneInterfaces.DroneEventsType;
 import org.droidplanner.core.drone.DroneInterfaces.OnDroneListener;
 import org.droidplanner.core.helpers.coordinates.Coord2D;
+import org.droidplanner.core.helpers.units.Altitude;
 import org.droidplanner.core.model.Drone;
+import org.droidplanner.core.survey.CameraInfo;
+import org.droidplanner.core.survey.Footprint;
+
+import com.MAVLink.Messages.ardupilotmega.msg_camera_feedback;
+import com.google.android.gms.internal.ln;
 
 import android.app.Activity;
 import android.content.Context;
@@ -25,6 +31,7 @@ import android.os.Bundle;
 import android.os.Handler;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -76,6 +83,8 @@ public abstract class DroneMap extends Fragment implements OnDroneListener {
 			}
 
 			mMapFragment.updateMissionPath(missionProxy);
+			
+			mMapFragment.updatePolygonsPaths(missionProxy.getPolygonsPath());
 
 			mHandler.removeCallbacks(this);
 		}
@@ -91,6 +100,8 @@ public abstract class DroneMap extends Fragment implements OnDroneListener {
 	public Drone drone;
 
 	protected Context context;
+
+	private CameraInfo camera = new CameraInfo();
 
 	protected abstract boolean isMissionDraggable();
 
@@ -179,6 +190,15 @@ public abstract class DroneMap extends Fragment implements OnDroneListener {
 			}
 			break;
 
+		case ATTITUDE:
+			if (((DroidPlannerApp) getActivity().getApplication()).getPreferences()
+					.isRealtimeFootprintsEnabled()) {
+				if (drone.getGps().isPositionValid()) {
+					mMapFragment.updateRealTimeFootprint(drone.getCamera().getCurrentFieldOfView());
+				}
+
+			}
+			break;
 		case GUIDEDPOINT:
 			mMapFragment.updateMarker(guided);
 			mMapFragment.updateDroneLeashPath(guided);
@@ -192,6 +212,9 @@ public abstract class DroneMap extends Fragment implements OnDroneListener {
 		case DISCONNECTED:
 		case HEARTBEAT_TIMEOUT:
 			mMapFragment.updateMarker(graphicDrone);
+			break;
+		case FOOTPRINT:
+				mMapFragment.addCameraFootprint(drone.getCamera().getLastFootprint());
 			break;
 		default:
 			break;