+++ Android/src/org/droidplanner/android/maps/DPMap.java
@@ -9,6 +9,7 @@ import org.droidplanner.android.utils.prefs.AutoPanMode;
 import org.droidplanner.core.drone.DroneInterfaces;
 import org.droidplanner.core.gcs.location.Location;
 import org.droidplanner.core.helpers.coordinates.Coord2D;
+import org.droidplanner.core.survey.Footprint;
 
 import android.graphics.Color;
 import android.location.LocationListener;
@@ -32,6 +33,13 @@ public interface DPMap extends DroneInterfaces.OnDroneListener {
 	public static final int DRONE_LEASH_DEFAULT_COLOR = Color.WHITE;
 	public static final int DRONE_LEASH_DEFAULT_WIDTH = 2;
 
+	public static final int POLYGONS_PATH_DEFAULT_COLOR = Color.RED;
+	public static final int POLYGONS_PATH_DEFAULT_WIDTH = 4;
+	
+	public static final int FOOTPRINT_DEFAULT_COLOR = 0;
+	public static final int FOOTPRINT_DEFAULT_WIDTH = 2;
+	public static final int FOOTPRINT_FILL_COLOR = Color.argb(80, 0, 0, 200);
+	
 	public static final String PREF_LAT = "pref_map_lat";
 	public static final float DEFAULT_LATITUDE = 37.8575523f;
 
@@ -133,6 +141,12 @@ public interface DPMap extends DroneInterfaces.OnDroneListener {
 	 *            drone's coordinate
 	 */
 	public void addFlightPathPoint(Coord2D coord);
+
+	/**
+	 * Draw the footprint of the camera in the ground
+	 * @param footprintToBeDraw
+	 */
+	public void addCameraFootprint(Footprint footprintToBeDraw);
 
 	/**
 	 * Remove all markers from the map.
@@ -343,6 +357,12 @@ public interface DPMap extends DroneInterfaces.OnDroneListener {
 	 *            source to use to draw the mission path
 	 */
 	public void updateMissionPath(PathSource pathSource);
+	
+	/**
+	 * Updates the polygons on the map.
+	 * 
+	 */
+    public void updatePolygonsPaths(List<List<Coord2D>> paths);
 
 	/**
 	 * Zoom to fit coordinates on map
@@ -363,5 +383,7 @@ public interface DPMap extends DroneInterfaces.OnDroneListener {
      * @param skip if it should skip further events
      */
     public void skipMarkerClickEvents(boolean skip);
+
+	public void updateRealTimeFootprint(Footprint footprint);
     
 }