+++ Android/src/org/droidplanner/android/maps/providers/google_map/GoogleMapFragment.java
@@ -21,6 +21,7 @@ import org.droidplanner.android.utils.collection.HashBiMap;
 import org.droidplanner.android.utils.prefs.AutoPanMode;
 import org.droidplanner.android.utils.prefs.DroidPlannerPrefs;
 import org.droidplanner.core.model.Drone;
+import org.droidplanner.core.survey.Footprint;
 import org.droidplanner.core.drone.DroneInterfaces;
 import org.droidplanner.core.helpers.coordinates.Coord2D;
 
@@ -54,6 +55,8 @@ import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.LatLngBounds;
 import com.google.android.gms.maps.model.Marker;
 import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.android.gms.maps.model.Polygon;
+import com.google.android.gms.maps.model.PolygonOptions;
 import com.google.android.gms.maps.model.Polyline;
 import com.google.android.gms.maps.model.PolylineOptions;
 import com.google.android.gms.maps.model.TileOverlay;
@@ -76,6 +79,7 @@ public class GoogleMapFragment extends SupportMapFragment implements DPMap, Loca
     private static final float USER_LOCATION_UPDATE_MIN_DISPLACEMENT = 15; // m
 
     private static final float GO_TO_MY_LOCATION_ZOOM = 19f;
+
 
     private final HashBiMap<MarkerInfo, Marker> mBiMarkersMap = new HashBiMap<MarkerInfo, Marker>();
 
@@ -109,6 +113,10 @@ public class GoogleMapFragment extends SupportMapFragment implements DPMap, Loca
 
 	protected boolean useMarkerClickAsMapClick = false;
     private boolean isMapLayoutFinished = false;
+
+	private List<Polygon> polygonsPaths = new ArrayList<Polygon>();
+
+	private Polygon footprintPoly;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup viewGroup,
@@ -516,8 +524,62 @@ public class GoogleMapFragment extends SupportMapFragment implements DPMap, Loca
 
         missionPath.setPoints(pathPoints);
     }
+    
+    
+    @Override
+	public void updateRealTimeFootprint(Footprint footprint) {
+    	if (footprintPoly == null) {
+    		PolygonOptions pathOptions = new PolygonOptions();
+    		pathOptions.strokeColor(FOOTPRINT_DEFAULT_COLOR).strokeWidth(FOOTPRINT_DEFAULT_WIDTH);
+    		pathOptions.fillColor(FOOTPRINT_FILL_COLOR);
 
-    /**
+    		for (Coord2D vertex : footprint.getVertexInGlobalFrame()) {
+    			pathOptions.add(DroneHelper.CoordToLatLang(vertex));
+    		}
+    		footprintPoly = getMap().addPolygon(pathOptions);
+    	}else{
+    		List<LatLng> list = new ArrayList<LatLng>();
+    		for (Coord2D vertex : footprint.getVertexInGlobalFrame()) {
+    			list.add(DroneHelper.CoordToLatLang(vertex));
+    		}
+    		footprintPoly.setPoints(list);
+    	}
+	}
+
+	@Override
+    public void updatePolygonsPaths(List<List<Coord2D>> paths){
+        for (Polygon poly : polygonsPaths) {
+			poly.remove();
+		}
+        
+        for (List<Coord2D> contour : paths) {
+        	PolygonOptions pathOptions = new PolygonOptions();
+            pathOptions.strokeColor(POLYGONS_PATH_DEFAULT_COLOR).strokeWidth(
+                    POLYGONS_PATH_DEFAULT_WIDTH);
+            final List<LatLng> pathPoints = new ArrayList<LatLng>(contour.size());
+			for (Coord2D coord : contour) {
+		            pathPoints.add(DroneHelper.CoordToLatLang(coord));
+			}
+			pathOptions.addAll(pathPoints);
+			polygonsPaths.add(getMap().addPolygon(pathOptions));
+		}
+        
+    }
+
+	@Override
+	public void addCameraFootprint(Footprint footprintToBeDraw) {
+		PolygonOptions pathOptions = new PolygonOptions();
+		pathOptions.strokeColor(FOOTPRINT_DEFAULT_COLOR).strokeWidth(FOOTPRINT_DEFAULT_WIDTH);
+		pathOptions.fillColor(FOOTPRINT_FILL_COLOR);
+
+		for (Coord2D vertex : footprintToBeDraw.getVertexInGlobalFrame()) {
+			pathOptions.add(DroneHelper.CoordToLatLang(vertex));
+		}
+		getMap().addPolygon(pathOptions);
+
+	}
+
+	/**
      * Save the map camera state on a preference file
      * http://stackoverflow.com/questions
      * /16697891/google-maps-android-api-v2-restoring