+++ Android/src/org/droidplanner/android/proxy/mission/item/fragments/MissionDetailFragment.java
@@ -107,6 +107,12 @@ public class MissionDetailFragment extends DialogFragment implements SpinnerSelf
 		case EPM_GRIPPER:
 			fragment = new MissionEpmGrabberFragment();
 			break;
+		case SET_SERVO:
+			fragment = new SetServoFragment();
+			break;
+		case CONDITION_YAW:
+			fragment = new MissionConditionYawFragment();
+			break;
 		default:
 			fragment = null;
 			break;