+++ Core/src/org/droidplanner/core/mission/Mission.java
@@ -14,8 +14,10 @@ import org.droidplanner.core.helpers.units.Length;
 import org.droidplanner.core.helpers.units.Speed;
 import org.droidplanner.core.mission.commands.CameraTrigger;
 import org.droidplanner.core.mission.commands.ChangeSpeed;
+import org.droidplanner.core.mission.commands.ConditionYaw;
 import org.droidplanner.core.mission.commands.EpmGripper;
 import org.droidplanner.core.mission.commands.ReturnToHome;
+import org.droidplanner.core.mission.commands.SetServo;
 import org.droidplanner.core.mission.commands.Takeoff;
 import org.droidplanner.core.mission.waypoints.Circle;
 import org.droidplanner.core.mission.waypoints.Land;
@@ -125,12 +127,13 @@ public class Mission extends DroneVariable {
 	 * @return the altitude of the last added mission item.
 	 */
 	public Altitude getLastAltitude() {
-		Altitude alt;
+		Altitude alt = defaultAlt;
 		try {
 			SpatialCoordItem lastItem = (SpatialCoordItem) items.get(items.size() - 1);
-			alt = lastItem.getCoordinate().getAltitude();
+			if (!(lastItem instanceof RegionOfInterest)) {
+				alt = lastItem.getCoordinate().getAltitude();
+			}
 		} catch (Exception e) {
-			alt = defaultAlt;
 		}
 		return alt;
 	}
@@ -257,6 +260,9 @@ public class Mission extends DroneVariable {
 
 		for (msg_mission_item msg : msgs) {
 			switch (msg.command) {
+			case MAV_CMD.MAV_CMD_DO_SET_SERVO:
+				received.add(new SetServo(msg, this));
+				break;
 			case MAV_CMD.MAV_CMD_NAV_WAYPOINT:
 				received.add(new Waypoint(msg, this));
 				break;
@@ -286,6 +292,10 @@ public class Mission extends DroneVariable {
 				break;
 			case MAV_CMD.MAV_CMD_NAV_RETURN_TO_LAUNCH:
 				received.add(new ReturnToHome(msg, this));
+				break;				
+			case MAV_CMD.MAV_CMD_CONDITION_YAW:
+				received.add(new ConditionYaw(msg, this));
+				break;
 			default:
 				break;
 			}